cloud_function/requirements.txt
google-cloud-bigquery==2.3.1
google-cloud-pubsub==2.1.0
gobits==0.0.7
